\name{kernelMatrix-methods}
\docType{methods}
\alias{kernelMatrix-methods}
\alias{kernelMatrix,stringkernelEx-method}
\alias{kernelMult-methods}
\alias{kernelMult,stringkernelEx-method}
\alias{kernelMult,stringkernelEx,ANY-method}
\alias{kernelPol-methods}
\alias{kernelPol,stringkernelEx-method}
\alias{show-methods}
\alias{show,stringkernelEx-method}
\alias{stringkernelEx-class}
\title{ Methods for stringKernelEx objects }
\description{
 S4 methods for kernels of the \code{stringkernelEx} class. 
 See \code{\link[kernlab]{kernelMatrix}} for details.
}
\keyword{methods}
\keyword{misc}
